Deduplicate the two API-key admin pages into a shared component
- analytics-web-app/src/lib/api-keys-shared.ts (new): factory providing the
  shared list/mint/revoke operations and error handling, parameterized by
  base path and the two response/error type differences
- analytics-web-app/src/components/ApiKeysAdminPage.tsx (new): shared page
  content (state, mint modal, one-time-key banner, revoke confirm, Next/
  Previous paging, table) parameterized by a config object
- analytics-web-app/src/lib/{analytics,ingestion}-api-keys-api.ts: now thin
  wrappers around the shared factory, keeping the same exported names
- analytics-web-app/src/routes/{Analytics,Ingestion}ApiKeysPage.tsx: now a
  module-level config constant plus a thin wrapper around the shared
  component, keeping their own Suspense/AuthGuard wrapping

No behavior change -- every documented difference (subtitle, placeholder,
revoke-confirm text, effective_within_seconds, optional error code, base
path, max-limit citation) is preserved in each page's own config. Existing
tests pass unmodified.
