+++ analytics-web-app/src/lib/analytics-api-keys-api.ts
@@ -1,14 +1,19 @@
 // Analytics API key management — calls `analytics-web-srv`'s own
 // `/api/analytics-api-keys` routes directly (no proxy needed: this service
-// hosts the analytics-key store itself). Modeled on `data-sources-api.ts`'s
-// `handleResponse`/error-class shape.
+// hosts the analytics-key store itself). Thin wrapper around the shared
+// factory in `api-keys-shared.ts` — see that file for the common
+// list/mint/revoke/error-handling logic.
 //
 // No `importKey` here: the import route exists for the `micromegas-import-keys`
 // CLI tool only (see the design doc's §5) — a browser form for pasting a
 // legacy key string in would reintroduce the "key transits a browser"
 // exposure mint already avoids.
 
-import { authenticatedFetch, getApiBase } from './api'
+import {
+  createApiKeysApi,
+  ApiKeyListEntry,
+  MintApiKeyResponse,
+} from './api-keys-shared'
 
 // The server's own hard cap (`MAX_LIMIT` in `rust/analytics-web-srv/src/analytics_keys.rs`),
 // used here as the page size for offset-based paging. Requested explicitly on
@@ -20,23 +25,9 @@ import { authenticatedFetch, getApiBase } from './api'
 // comparing the returned row count against this same value.
 export const MAX_ANALYTICS_API_KEYS_LIST_LIMIT = 500
 
-export interface AnalyticsApiKeyListEntry {
-  key_id: string
-  name: string
-  created_at: string
-  created_by: string
-  last_used_at: string | null
-  revoked_at: string | null
-  revoked_by: string | null
-}
+export type AnalyticsApiKeyListEntry = ApiKeyListEntry
 
-export interface MintAnalyticsApiKeyResponse {
-  key_id: string
-  name: string
-  created_at: string
-  /** The cleartext key, returned exactly once. Never persisted client-side. */
-  key: string
-}
+export type MintAnalyticsApiKeyResponse = MintApiKeyResponse
 
 export interface RevokeAnalyticsApiKeyResponse {
   revoked_at: string
@@ -47,59 +38,14 @@ export interface AnalyticsApiKeyErrorResponse {
   message: string
 }
 
-export class AnalyticsApiKeyError extends Error {
-  constructor(
-    public code: string,
-    message: string,
-    public status: number
-  ) {
-    super(message)
-    this.name = 'AnalyticsApiKeyError'
-  }
-}
+const api = createApiKeysApi<AnalyticsApiKeyErrorResponse, RevokeAnalyticsApiKeyResponse>({
+  basePath: '/analytics-api-keys',
+  errorName: 'AnalyticsApiKeyError',
+  maxListLimit: MAX_ANALYTICS_API_KEYS_LIST_LIMIT,
+})
 
-async function handleResponse<T>(response: Response): Promise<T> {
-  if (!response.ok) {
-    let errorData: AnalyticsApiKeyErrorResponse | undefined
-    try {
-      errorData = await response.json()
-    } catch {
-      // Ignore JSON parse errors
-    }
-    throw new AnalyticsApiKeyError(
-      errorData?.code ?? 'UNKNOWN_ERROR',
-      errorData?.message ?? `HTTP ${response.status}`,
-      response.status
-    )
-  }
-  return response.json()
-}
+export const AnalyticsApiKeyError = api.ErrorClass
 
-export async function listAnalyticsApiKeys(
-  includeRevoked = true,
-  offset = 0
-): Promise<AnalyticsApiKeyListEntry[]> {
-  const response = await authenticatedFetch(
-    `${getApiBase()}/analytics-api-keys?limit=${MAX_ANALYTICS_API_KEYS_LIST_LIMIT}&offset=${offset}&include_revoked=${includeRevoked}`
-  )
-  return handleResponse<AnalyticsApiKeyListEntry[]>(response)
-}
-
-export async function mintAnalyticsApiKey(name: string): Promise<MintAnalyticsApiKeyResponse> {
-  const response = await authenticatedFetch(`${getApiBase()}/analytics-api-keys`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ name }),
-  })
-  return handleResponse<MintAnalyticsApiKeyResponse>(response)
-}
-
-export async function revokeAnalyticsApiKey(
-  keyId: string
-): Promise<RevokeAnalyticsApiKeyResponse> {
-  const response = await authenticatedFetch(
-    `${getApiBase()}/analytics-api-keys/${encodeURIComponent(keyId)}`,
-    { method: 'DELETE' }
-  )
-  return handleResponse<RevokeAnalyticsApiKeyResponse>(response)
-}
+export const listAnalyticsApiKeys = api.list
+export const mintAnalyticsApiKey = api.mint
+export const revokeAnalyticsApiKey = api.revoke
